requirements.txt
fastapi
uvicorn[standard]
requests
msal
azure-storage-blob
python-dotenv
